+++ benches/bench.rs
@@ -1,7 +1,9 @@
-//! Currently not used, because benchmarking is nightly only.
+#![feature(test)]
+//! Note that benchmarks can easily be misleading.
+//! The Alternating algorithm comes out fastest - but it finds much worse solutions.
 extern crate test;
 
-use super::*;
+use kmedoids::*;
 use ndarray::Array2;
 use rand::{rngs::StdRng, Rng, SeedableRng};
 use test::{black_box, Bencher};
@@ -99,6 +101,26 @@ fn bench_pam(b: &mut Bencher) {
 	}
 	b.iter(|| {
 		let (loss, assignment, _, _, _) = pam(&mat, 5, 100);
+		black_box(loss);
+		black_box(assignment);
+	});
+}
+
+#[bench]
+fn bench_alternating(b: &mut Bencher) {
+	let n = 100;
+	let mut rng = StdRng::seed_from_u64(42);
+	let mut mat = Array2::<i32>::from_elem((n, n), 0);
+	for i in 0..n {
+		for j in (i + 1)..n {
+			let v = rng.gen_range(1..100);
+			mat[[i, j]] = v;
+			mat[[j, i]] = v;
+		}
+	}
+	b.iter(|| {
+		let mut med = vec![0, 1, 2, 3, 4];
+		let (loss, assignment, _) = alternating(&mat, &mut med, 100);
 		black_box(loss);
 		black_box(assignment);
 	});